requirements.txt
altgraph==0.17.2
blis==0.7.4
catalogue==2.0.6
certifi==2021.5.30
charset-normalizer==2.0.5
click==7.1.2
colorama==0.4.4
cymem==2.0.5
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.1.0/en_core_web_sm-3.1.0-py3-none-any.whl
future==0.18.2
idna==3.2
Jinja2==3.0.1
jmespath==1.0.0
MarkupSafe==2.0.1
murmurhash==1.0.5
ngrok==0.0.1
numpy==1.21.2
packaging==21.0
pandas==1.4.2
pathy==0.6.0
pefile==2021.9.3
preshed==3.0.5
psycopg2==2.9.3
pydantic==1.8.2
pyinstaller==4.10
pyinstaller-hooks-contrib==2022.3
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32-ctypes==0.2.0
requests==2.26.0
six==1.16.0
smart-open==5.2.1
spacy==3.1.2
spacy-legacy==3.0.8
srsly==2.4.1
thinc==8.0.10
tqdm==4.62.2
twilio==7.8.1
typer==0.3.2
typing-extensions==3.10.0.2
urllib3==1.26.6
wasabi==0.8.2
